Add explicit type conversions where necessary: storage/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2945968
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890884}

+++ storage/browser/blob/blob_impl.cc
@@ -231,7 +231,8 @@ void BlobImpl::CaptureSnapshot(CaptureSnapshotCallback callback) {
                   base::File::Info info;
                   if (!base::GetFileInfo(path, &info))
                     return SizeAndTime{0, absl::nullopt};
-                  return SizeAndTime{info.size, info.last_modified};
+                  return SizeAndTime{static_cast<uint64_t>(info.size),
+                                     info.last_modified};
                 },
                 item->path()),
             base::BindOnce(
